Move annotations to new location
