+++ src/main/java/com/mealtiger/backend/configuration/ConfigLoader.java
@@ -1,5 +1,6 @@
 package com.mealtiger.backend.configuration;
 
+import com.mealtiger.backend.configuration.annotations.Config;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.yaml.snakeyaml.DumperOptions;
