+++ src/main/java/com/mealtiger/backend/configuration/annotations/Config.java
@@ -1,4 +1,4 @@
-package com.mealtiger.backend.configuration;
+package com.mealtiger.backend.configuration.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
