+++ src/main/java/com/mealtiger/backend/configuration/annotations/ConfigNode.java
@@ -1,4 +1,4 @@
-package com.mealtiger.backend.configuration;
+package com.mealtiger.backend.configuration.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -18,4 +18,9 @@ public @interface ConfigNode {
      * Represents a property from a config file. Example: Database.mongoDBURL
      */
     String name();
+
+    /**
+     * If changed an environment variable of this name will override the config setting
+     */
+    String envKey() default "";
 }
